Add TaskAssignmentMixin and move to mixins file
